Cambiato da int a Integer e rimosse configurazioni di connessione al db.

+++ exprv/src/main/java/com/prenotazioni/exprivia/exprv/controller/UserController.java
@@ -37,7 +37,7 @@ public class UserController {
 
     // Gestisce le richieste GET per ottenere un singolo utente tramite ID
     @GetMapping("/utente")
-    public Users getUtente(@PathVariable Long id) {
+    public Users getUtente(@PathVariable Integer id) {
         return userService.cercaSingolo(id); // Chiama il servizio per ottenere un utente specifico
     }
 
@@ -55,7 +55,7 @@ public class UserController {
 
     // Gestisce le richieste PUT per aggiornare un utente esistente tramite ID
     @PutMapping("/{id}")
-    public ResponseEntity<?> aggiornaUser(@PathVariable Long id, @RequestBody Users user) {
+    public ResponseEntity<?> aggiornaUser(@PathVariable Integer id, @RequestBody Users user) {
         try {
             Users userAggiornato = userService.aggiornaUser(id, user);// Chiama il servizio per aggiornare l'utente
             return ResponseEntity.ok(userAggiornato);// Restituisce una risposta con stato 200 (OK) e l'utente aggiornato
@@ -66,7 +66,7 @@ public class UserController {
 
     // Gestisce le richieste DELETE per eliminare un utente tramite ID
     @DeleteMapping("/{id}")
-    public ResponseEntity<String> eliminaUser(@PathVariable Long id) {
+    public ResponseEntity<String> eliminaUser(@PathVariable Integer id) {
         try {
             userService.eliminaUser(id);// Chiama il servizio per eliminare l'utente
             return ResponseEntity.noContent().build(); // Restituisce una risposta con stato 204 (NO CONTENT)
